SYNERGY-885 mac client listens on ipv4 only but attempts to connect on ipv6 *Fix code smells

+++ src/lib/synergy/ClientApp.h
@@ -84,5 +84,5 @@ private:
     Client*            m_client;
     synergy::Screen*   m_clientScreen;
     NetworkAddress*    m_serverAddress;
-    int                m_lastServerAddressIndex = 0;
+    size_t             m_lastServerAddressIndex = 0;
 };
